refactor: Apply code formatting using clang-format all native code (#733)
* apply code format

* fix build error

* refactor

* fix

* fix

* fix

* fix

+++ Plugin~/WebRTCPlugin/AudioTrackSinkAdapter.cpp
@@ -1,7 +1,9 @@
 #include "pch.h"
+
+#include <audio/remix_resample.h>
+#include <common_audio/include/audio_util.h>
+
 #include "AudioTrackSinkAdapter.h"
-#include "audio/remix_resample.h"
-#include "common_audio/include/audio_util.h"
 
 namespace unity
 {
@@ -12,14 +14,14 @@ namespace webrtc
     {
     }
 
-    AudioTrackSinkAdapter::~AudioTrackSinkAdapter()
-    {
-        WebRtc_FreeBuffer(_buffer);
-    }
+    AudioTrackSinkAdapter::~AudioTrackSinkAdapter() { WebRtc_FreeBuffer(_buffer); }
 
     void AudioTrackSinkAdapter::OnData(
-        const void* audio_data, int bits_per_sample, int sample_rate,
-        size_t number_of_channels, size_t number_of_frames)
+        const void* audio_data,
+        int bits_per_sample,
+        int sample_rate,
+        size_t number_of_channels,
+        size_t number_of_frames)
     {
         std::lock_guard<std::mutex> lock(_mutex);
 
@@ -40,22 +42,19 @@ namespace webrtc
             &_resampler,
             &_frame);
 
-        size_t length =
-            _frame.num_channels() *
-            _frame.samples_per_channel();
+        size_t length = _frame.num_channels() * _frame.samples_per_channel();
 
         WebRtc_WriteBuffer(_buffer, _frame.data(), length);
     }
 
-    void AudioTrackSinkAdapter::ResizeBuffer(
-        size_t channels, int sampleRate, size_t length)
+    void AudioTrackSinkAdapter::ResizeBuffer(size_t channels, int32_t sampleRate, size_t length)
     {
         RTC_DCHECK(channels);
         RTC_DCHECK(sampleRate);
 
         // reallocate ring buffer, keep it relatively short at 0.2s
-        size_t bufferSize = static_cast<size_t>(channels * sampleRate * 0.2f);
-        if(_buffer != nullptr)
+        size_t bufferSize = static_cast<size_t>(static_cast<float>(channels) * static_cast<float>(sampleRate) * 0.2f);
+        if (_buffer != nullptr)
             WebRtc_FreeBuffer(_buffer);
         _buffer = WebRtc_CreateBuffer(bufferSize, sizeof(int16_t));
 
@@ -67,8 +66,7 @@ namespace webrtc
         _bufferIn.resize(length);
     }
 
-    void AudioTrackSinkAdapter::ProcessAudio(
-        float* data, size_t length, size_t channels, int32_t sampleRate)
+    void AudioTrackSinkAdapter::ProcessAudio(float* data, size_t length, size_t channels, int32_t sampleRate)
     {
         RTC_DCHECK(data);
         RTC_DCHECK(length);
@@ -81,16 +79,13 @@ namespace webrtc
 
         // Reallocate audio buffer when Unity changes channel count, sample rate,
         // or data length.
-        if (_buffer == nullptr ||
-            _frame.num_channels_ != channels ||
-            _frame.sample_rate_hz_ != sampleRate ||
+        if (_buffer == nullptr || _frame.num_channels_ != channels || _frame.sample_rate_hz_ != sampleRate ||
             _bufferIn.size() != length)
         {
             ResizeBuffer(channels, sampleRate, length);
         }
 
-        size_t readLength = WebRtc_ReadBuffer(
-            _buffer, nullptr, _bufferIn.data(), length);
+        size_t readLength = WebRtc_ReadBuffer(_buffer, nullptr, _bufferIn.data(), length);
 
         for (size_t i = 0; i < readLength; i++)
             data[i] = webrtc::S16ToFloat(_bufferIn[i]);
